pyproject.toml
[tool.poetry]
name = "shortn"
version = "1.0.0"
description = "A reusable URL shortening microservice built with Django REST Framework"
authors = ["Camryn Crowder <camryn.crowder@apkudo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
qrcode = "^8.0"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.4.0"
python-decouple = "^3.8"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
gunicorn = "^23.0.0"
whitenoise = "^6.7.0"
django-redis = "^5.4.0"
confluent-kafka = "^2.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.8"

[virtualenvs]
create = true
in-project = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
